Time: 3 ms (38.99%), Space: 7.1 MB (65.45%) - LeetHub

+++ 1922-count-good-numbers/1922-count-good-numbers.cpp
@@ -3,9 +3,9 @@ public:
     const long long   m = 1e9+7;
 
     long long power(long long x , long long y){
-        if(y==0){
-            return 1;
-        }
+       if(y==0){
+        return 1;
+       }
        long long temp = power(x,y/2);
        temp = (temp*temp);
 
@@ -17,13 +17,11 @@ public:
        }
 
     }
-
     int countGoodNumbers(long long n) {
-
         long long  pow1=n/2+n%2;
         long long  pow2 = n/2;
         long long  ans1 = power(5,pow1);
         long long  ans2 = power(4,pow2);
-        return (ans1*ans2)%m;
+        return int((ans1*ans2)%m);
     }
 };
